+++ src/gdik-select/GDIKSelect.js
@@ -124,7 +124,7 @@ export default class GDIKSelect extends HTMLElement {
             }
             this.mapElement.appendChild(this.searchElement);
         }
-        this.dispatchEvent(new CustomEvent("onConfigLoaded", e));
+        this.dispatchEvent(new CustomEvent("configloaded", e));
     }
 
     setValue (value, silent) {